+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:dooadex_constants/dooadex_constants.dart';
+import 'package:dooadex_logger/dooadex_logger.dart';
 import 'package:dooadex_palette/dooadex_palette.dart';
 import 'package:flutter/material.dart';
 
@@ -13,6 +14,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       title: 'Dooadex Package Test App',
       theme: ThemeData(
         primarySwatch: MaterialColor(DooadexColor.primaryMaterialColor.colorHex, DooadexColor.primaryMaterialColor.swatch),
@@ -40,18 +42,15 @@ class _MyHomePageState extends State<MyHomePage> {
       body: Center(
         child: ElevatedButton(
           onPressed: () {
-            // DooadexRoute.init(routePath: DooadexRoute.path, namedRoute: DooadexRoute.name);
-            // print(DooadexRoute.path.home);
-            // print(DooadexRoute.name.home);
-            print(HttpStatus.code.badRequest);
-            print(HttpStatus.reason.badRequest);
-            print(HttpConstant.method.get);
-            print(HttpConstant.method.post);
-            print(HttpConstant.method.put);
-            print(HttpConstant.method.patch);
-            print(HttpConstant.method.delete);
-            print(HttpConstant.contentType.json);
-            print(HttpConstant.tokenType.bearer);
+            DdxLogger(DdxHttpStatus.code.badRequest);
+            DdxLogger(DdxHttpStatus.reason.badRequest);
+            DdxLogger(DdxHttpConst.method.get);
+            DdxLogger(DdxHttpConst.method.post);
+            DdxLogger(DdxHttpConst.method.put);
+            DdxLogger(DdxHttpConst.method.patch);
+            DdxLogger(DdxHttpConst.method.delete);
+            DdxLogger(DdxHttpConst.contentType.json);
+            DdxLogger(DdxHttpConst.tokenType.bearer);
           },
           child: const Text("Test"),
         ),