+++ lib/src/http/http_status.dart
@@ -1,8 +1,8 @@
 import 'http_status_code.dart';
 import 'http_status_reason.dart';
 
-class HttpStatus {
-  HttpStatus._();
+class DdxHttpStatus {
+  DdxHttpStatus._();
 
   static HttpStatusCode code = HttpStatusCode();
   static HttpStatusReason reason = HttpStatusReason();