+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.4
 
-Remove route constant
+Changed Variable prefix Dooadex to Ddx.